backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.10.3
python-multipart==0.0.6
google-generativeai>=0.8.0,<1.0.0
pydantic>=2.7.0,<3.0.0
python-dotenv==1.0.0
httpx==0.25.2
opencv-python-headless>=4.10.0,<5.0.0
numpy>=2.0.0,<3.0.0
pillow>=10.0.0
nltk==3.8.1


